[perf-] Allow filling the buffer from outside parser
This will allow the buffer to be filled synchronously or
asynchronously, allowing the overall parser to be used in more
contexts.

To avoid borrow-checker limitations, we have to return a `usize` back
out of the core parser instead of a `&str`.

We also have to be able to rollback some state if we run out of input
in the middle of a parse. This means that we could be redoing some
work occasionally.

This has a small performance loss (111MB XML file / 8983678
tokens):

Before:

```
  Time (mean ± σ):      2.653 s ±  0.017 s    [User: 2.634 s, System: 0.019 s]
  Range (min … max):    2.638 s …  2.698 s    10 runs
```

After:

```
  Time (mean ± σ):      2.690 s ±  0.014 s    [User: 2.674 s, System: 0.016 s]
  Range (min … max):    2.666 s …  2.712 s    10 runs
```

That's 101.39% of the previous time.

```
  Instructions:          5343004592 (+10.03603%)
  L1 Accesses:           6929446615 (+12.81603%)
  Estimated Cycles:      7222711720 (+12.23203%)
```